fixes bug where I was always grabbing the basic table for multi-table pages.
this bug applied to both the player career stats pages and the playoff stats pages,
and unfortunately went pretty deep -- it was happening because the other tables are
commented out in the page.  to handle that, I switched to beautifulsoup (via b4),
search the entire document for any reasonable tables (commented out or otherwise),
and then return a dictionary of those tables for the parser to grab.  slightly
changes the text parsing as well.  this commit also includes many small fixes to
datatypes due to overlapping column names between per X and total tables --
i was parsing those as integers (for the totals), but the per X tables are obviously
floats.  now they're named differently.